VirusGeeks/Billing_Report/requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.11.1
boto3==1.22.11
botocore==1.25.11
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==37.0.2
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.2.1
fernet==1.0.1
Flask==2.1.1
Flask-Compress==1.11
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
msal==1.17.0
numpy==1.22.3
Office365-REST-Python-Client==2.3.11
packaging==21.3
pandas==1.4.2
plotly==5.6.0
pyaes==1.6.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redshift-connector==2.0.907
requests==2.27.1
s3transfer==0.5.2
scramp==1.4.1
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.0.1
urllib3==1.26.9
Werkzeug==2.1.1
XlsxWriter==3.0.3